+++ eqcorrscan/tests/mag_calc_test.py
@@ -24,7 +24,7 @@ from obspy.io.nordic.core import readwavename
 from eqcorrscan.utils import mag_calc
 from eqcorrscan.utils.mag_calc import (
     dist_calc, _sim_WA, _max_p2t, _GSE2_PAZ_read, _find_resp, _pairwise,
-    svd_moments, amp_pick_event, _snr, relative_amplitude)
+    svd_moments, amp_pick_event, _snr, relative_amplitude, relative_magnitude)
 from eqcorrscan.utils.clustering import svd
 from eqcorrscan.helpers.mock_logger import MockLoggingHandler
 
@@ -218,6 +218,23 @@ class TestMagCalcMethods(unittest.TestCase):
 
 class TestRelativeAmplitudes(unittest.TestCase):
     """Test the relative amplitude methods."""
+    @classmethod
+    @pytest.mark.network
+    def setUpClass(cls):
+        client = Client("GEONET")
+        cls.event1 = client.get_events(eventid="2016p912302")[0]
+        cls.event2 = client.get_events(eventid="3470170")[0]
+        bulk = [(p.waveform_id.network_code, p.waveform_id.station_code,
+                 p.waveform_id.location_code, p.waveform_id.channel_code,
+                 p.time - 20, p.time + 60) for p in cls.event1.picks]
+        st1 = client.get_waveforms_bulk(bulk)
+        cls.st1 = st1.detrend().filter("bandpass", freqmin=2, freqmax=20)
+        bulk = [(p.waveform_id.network_code, p.waveform_id.station_code,
+                 p.waveform_id.location_code, p.waveform_id.channel_code,
+                 p.time - 20, p.time + 60) for p in cls.event2.picks]
+        st2 = client.get_waveforms_bulk(bulk)
+        cls.st2 = st2.detrend().filter("bandpass", freqmin=2, freqmax=20)
+
     def test_snr(self):
         noise = np.random.randn(100)
         signal = np.random.randn(100)
@@ -335,25 +352,31 @@ class TestRelativeAmplitudes(unittest.TestCase):
         for value in relative_amplitudes.values():
             self.assertAlmostEqual(value, scale_factor)
 
-    @pytest.mark.network
     def test_real_near_repeat(self):
-        client = Client("GEONET")
-        event1 = client.get_events(eventid="2016p912302")[0]
-        event2 = client.get_events(eventid="3470170")[0]
-        bulk = [(p.waveform_id.network_code, p.waveform_id.station_code,
-                 p.waveform_id.location_code, p.waveform_id.channel_code,
-                 p.time - 20, p.time + 60) for p in event1.picks]
-        st1 = client.get_waveforms_bulk(bulk)
-        st1.detrend().filter("bandpass", freqmin=2, freqmax=20)
-        bulk = [(p.waveform_id.network_code, p.waveform_id.station_code,
-                 p.waveform_id.location_code, p.waveform_id.channel_code,
-                 p.time - 20, p.time + 60) for p in event2.picks]
-        st2 = client.get_waveforms_bulk(bulk)
-        st2.detrend().filter("bandpass", freqmin=2, freqmax=20)
         relative_amplitudes = relative_amplitude(
-            st1=st1, st2=st2, event1=event1, event2=event2)
+            st1=self.st1, st2=self.st2, event1=self.event1, event2=self.event2)
         for seed_id, ratio in relative_amplitudes.items():
             self.assertLess(abs(0.8 - ratio), 0.1)
+
+    def test_real_near_repeat_magnitudes(self):
+        relative_magnitudes, correlations = relative_magnitude(
+            st1=self.st1, st2=self.st2, event1=self.event1, event2=self.event2,
+            return_correlations=True)
+        for seed_id, mag_diff in relative_magnitudes.items():
+            self.assertLess(abs(mag_diff + 0.15), 0.1)
+
+    def test_real_near_repeat_magnitudes_corr_provided(self):
+        relative_magnitudes = relative_magnitude(
+            st1=self.st1, st2=self.st2, event1=self.event1, event2=self.event2,
+            correlations={tr.id: 1.0 for tr in self.st1})
+        for seed_id, mag_diff in relative_magnitudes.items():
+            self.assertLess(abs(mag_diff + 0.15), 0.1)
+
+    def test_real_near_repeat_magnitudes_bad_corr(self):
+        relative_magnitudes = relative_magnitude(
+            st1=self.st1, st2=self.st2, event1=self.event1, event2=self.event2,
+            correlations={tr.id: 0.2 for tr in self.st1})
+        self.assertEqual(len(relative_magnitudes), 0)
 
 
 class TestAmpPickEvent(unittest.TestCase):